Added a moon to the skybox
The moon is always located opposite to the sun
