+++ source/client/assets/shaders/shader_celestial_body.frag
