+++ source/client/source/Constants.cs
@@ -46,7 +46,7 @@ public static class Constants
     public const int CHUNK_COLUMN_HEIGHT_BLOCKS = CHUNK_COLUMN_HEIGHT * CHUNK_SIZE;
     
     public const bool CIRCULAR_LOAD_REGION = true;
-    public const int CHUNK_COLUMN_LOAD_RADIUS = 32;
+    public const int CHUNK_COLUMN_LOAD_RADIUS = 12;
     public const int CHUNK_COLUMN_LOAD_RADIUS_SQUARED = CHUNK_COLUMN_LOAD_RADIUS * CHUNK_COLUMN_LOAD_RADIUS;
     public const int CHUNK_COLUMN_UNLOAD_RADIUS = CHUNK_COLUMN_LOAD_RADIUS + 3;
     public const int CHUNK_COLUMN_UNLOAD_RADIUS_SQUARED = CHUNK_COLUMN_UNLOAD_RADIUS * CHUNK_COLUMN_UNLOAD_RADIUS;
