+++ source/client/source/Rendering/Shaders/ShaderManager.cs
@@ -10,7 +10,7 @@ public class ShaderManager : IDisposable
     public static Shader DebugShader { get; private set; } = null!;
     public static Shader ChunkShader { get; private set; } = null!;
     public static Shader SkyboxShader { get; private set; } = null!;
-    public static Shader SunShader { get; private set; } = null!;
+    public static Shader CelestialBodyShader { get; private set; } = null!;
     
     public static Matrix4 ProjectionMatrix { get; private set; } = Matrix4.Identity;
     public static Matrix4 ViewMatrix { get; private set; } = Matrix4.Identity;
@@ -30,8 +30,8 @@ public class ShaderManager : IDisposable
         SkyboxShader = new Shader(IoUtils.GetShaderPath("shader_skybox.vert"), IoUtils.GetShaderPath("shader_skybox.frag"));
         SkyboxShader.Use();
         
-        SunShader = new Shader(IoUtils.GetShaderPath("shader_sun.vert"), IoUtils.GetShaderPath("shader_sun.frag"));
-        SunShader.Use();
+        CelestialBodyShader = new Shader(IoUtils.GetShaderPath("shader_celestial_body.vert"), IoUtils.GetShaderPath("shader_celestial_body.frag"));
+        CelestialBodyShader.Use();
     }
     
     
@@ -51,8 +51,8 @@ public class ShaderManager : IDisposable
         SkyboxShader.Use();
         SkyboxShader.SetMatrix4("projection", projectionMatrix);
         
-        SunShader.Use();
-        SunShader.SetMatrix4("projection", projectionMatrix);
+        CelestialBodyShader.Use();
+        CelestialBodyShader.SetMatrix4("projection", projectionMatrix);
     }
     
     
@@ -99,7 +99,7 @@ public class ShaderManager : IDisposable
         DebugShader.Dispose();
         ChunkShader.Dispose();
         SkyboxShader.Dispose();
-        SunShader.Dispose();
+        CelestialBodyShader.Dispose();
         
         GC.SuppressFinalize(this);
     }
