+++ source/client/source/Rendering/Skybox/Moon.cs
@@ -1,0 +1,24 @@
+﻿using Korpi.Client.Utils;
+using Korpi.Client.Window;
+using OpenTK.Mathematics;
+
+namespace Korpi.Client.Rendering.Skybox;
+
+public class Moon : CelestialBody
+{
+    public Moon(bool enableRotation) : base(enableRotation, new[]
+    {
+        IoUtils.GetSkyboxSunTexturePath("moon.png"),
+        IoUtils.GetSkyboxSunTexturePath("moon.png"),
+        IoUtils.GetSkyboxSunTexturePath("moon.png"),
+        IoUtils.GetSkyboxSunTexturePath("moon.png"),
+        IoUtils.GetSkyboxSunTexturePath("moon.png"),
+        IoUtils.GetSkyboxSunTexturePath("moon.png"),
+    }, 5)
+    {
+    }
+
+
+    protected override Vector3 Position => -GameTime.SunDirection * 800; // WARN: The proper way would be to use a orthographic projection matrix
+    protected override float Scale => 40f;
+}
